Struct libraymarcher::RayMarcherConfig

pub struct RayMarcherConfig<C> {
    pub width: usize,
    pub height: usize,
    pub camera_pos: Vec3,
    pub look_at: Vec3,
    pub light_pos: Vec3,
    pub background_color: C,
    pub camera_zoom: f64,
    pub anti_aliasing_level: u32,
    pub backplane_positions: Vec3,
    pub specular_shininess: f64,
    pub specular_color: C,
}

Fields
width: usize

Width of the rendered image in px

height: usize

Height of the rendered image in px

camera_pos: Vec3

Camera position in 3d space

look_at: Vec3

Point that the camera should point towards (usually leave this at (0, 0, 0) for julia sets)

light_pos: Vec3

Position of the Phong directional light in 3d space

background_color: C

Color to render if a ray missed the scene

camera_zoom: f64

Zoom level of the camera. 1.0 normal zoom

anti_aliasing_level: u32

Size of subpixel grid. anti_aliasing_level of 4 will create a 4x4 subpixel grid (so 16x AA)

backplane_positions: Vec3

Position of the back culling planes. Any rays that hit this will be assumed to be a miss

specular_shininess: f64

Phong shininess constant

specular_color: C

Color of specular highlights