Struct libraymarcher::Vec3

pub struct Vec3 {
    pub x: f64,
    pub y: f64,
    pub z: f64,
}

A 3 dimensional vector of f64

Fields
x: f64y: f64z: f64
Implementations

impl Vec3

pub fn magnitude(self) -> f64

The Euclidean norm

pub fn dot(self, other: Vec3) -> f64

Dot product with another Vec3

pub fn normalized(self) -> Self

Normalize this Vec3 so that it has a magnitude of 1

Will panic with a div by zero if the self.magnitude() is zero

pub fn reflect(self, norm: Vec3) -> Self

Reflect this vector about a normal vec.

pub fn cross(self, other: Vec3) -> Self

Cross product with another Vec3

pub fn refract(self, norm: Vec3, eta: f64) -> Self

Simulate Snell's law refrection through a surface with normal vec.

eta is the relative index of refraction, ri1 / ri2

pub fn lerp(self, other: Vec3, t: f64) -> Self

Linearly interpolate between two Vec3 with an interpolation value t